+++ tests/testthat/test-tTOST.R
@@ -41,10 +41,12 @@ test_that("Run examples for one sample", {
                  low_eqbound = -.5,
                  high_eqbound = .5)
   test1_smd = smd_calc(x = samp1,
-                       alpha = .1)
+                       alpha = .1,
+                       output = "data.frame")
   test1_smd_boot = boot_smd_calc(x = samp1,
                        alpha = .1,
-                       R = 99)
+                       R = 99,
+                       output = "data.frame")
   expect_equal(test1_smd$estimate, test1_smd_boot$estimate)
   expect_equal(test1_smd$estimate, test1$effsize$estimate[2])
   expect_equal(test1_smd$lower.ci, test1$effsize$lower.ci[2])
@@ -304,18 +306,21 @@ test_that("Run examples for two sample", {
 
   test1_smd = smd_calc(x = samp1,
                               y = samp2,
-                       alpha = .1)
+                       alpha = .1,
+                       output = "data.frame")
 
   test1_smd_boot_stud = boot_smd_calc(x = samp1,
                        y = samp2,
                        alpha = .1,
                        boot_ci = "stud",
-                       R = 99)
+                       R = 99,
+                       output = "data.frame")
   test1_smd_boot_basic = boot_smd_calc(x = samp1,
                                       y = samp2,
                                       alpha = .1,
                                       boot_ci = "b",
-                                      R = 99)
+                                      R = 99,
+                                      output = "data.frame")
   expect_error(boot_smd_calc(x = samp1,
                              y = samp2,
                              alpha = .1,
@@ -325,7 +330,8 @@ test_that("Run examples for two sample", {
                                       y = samp2,
                                       alpha = .1,
                                       boot_ci = "p",
-                                      R = 99)
+                                      R = 99,
+                                      output = "data.frame")
   expect_equal(rep(test1_smd$estimate,3),
                c(test1_smd_boot_stud$estimate,
                  test1_smd_boot_basic$estimate,
@@ -555,21 +561,24 @@ test_that("Run examples for two sample", {
   test1_smd = smd_calc(formula = y ~ group,
                        data = df_samp,
                        var.equal = TRUE,
-                       bias_correction = FALSE)
+                       bias_correction = FALSE,
+                       output = "data.frame")
   test1_smd_boot_stud = boot_smd_calc(x = samp1,
                                       y = samp2,
                                       alpha = .1,
                                       boot_ci = "stud",
                                       R = 99,
                                       var.equal = TRUE,
-                                      bias_correction = FALSE)
+                                      bias_correction = FALSE,
+                                      output = "data.frame")
   test1_smd_boot_basic = boot_smd_calc(x = samp1,
                                        y = samp2,
                                        alpha = .1,
                                        boot_ci = "b",
                                        R = 99,
                                        var.equal = TRUE,
-                                       bias_correction = FALSE)
+                                       bias_correction = FALSE,
+                                       output = "data.frame")
   expect_error(boot_smd_calc(x = samp1,
                              y = samp2,
                              alpha = .1,
@@ -583,7 +592,8 @@ test_that("Run examples for two sample", {
                                       boot_ci = "p",
                                       R = 99,
                                       var.equal = TRUE,
-                                      bias_correction = FALSE)
+                                      bias_correction = FALSE,
+                                      output = "data.frame")
   expect_equal(rep(test1_smd$estimate,3),
                c(test1_smd_boot_stud$estimate,
                  test1_smd_boot_basic$estimate,
@@ -676,18 +686,21 @@ test_that("Run examples for paired samples", {
   test1_smd = smd_calc(x = samp1,
                  y = samp2,
                  paired = TRUE,
-                 alpha = .1)
+                 alpha = .1,
+                 output = "data.frame")
   test1_smd_boot_stud = boot_smd_calc(x = samp1,
                                       y = samp2,
                                       paired = TRUE,
                                       alpha = .1,
-                                      R = 99)
+                                      R = 99,
+                                      output = "data.frame")
   test1_smd_boot_basic = boot_smd_calc(x = samp1,
                                        y = samp2,
                                        paired = TRUE,
                                        alpha = .1,
                                        boot_ci = "b",
-                                       R = 99)
+                                       R = 99,
+                                       output = "data.frame")
   expect_error(boot_smd_calc(x = samp1,
                              y = samp2,
                              alpha = .1,
@@ -701,7 +714,8 @@ test_that("Run examples for paired samples", {
                                       paired = TRUE,
                                       alpha = .1,
                                       boot_ci = "p",
-                                      R = 99)
+                                      R = 99,
+                                      output = "data.frame")
   expect_equal(rep(test1_smd$estimate,3),
                c(test1_smd_boot_stud$estimate,
                  test1_smd_boot_basic$estimate,